@@ -3,7 +3,7 @@
     "type": "library",
     "description": "Pico is a flat file CMS, this means there is no administration backend and database to deal with. You simply create .md files in the \"content\" folder and that becomes a page.",
     "keywords": ["pico", "picocms", "pico-cms", "simple", "flat-file", "cms", "content-management", "website", "markdown-to-html", "php", "markdown", "yaml", "twig" ],
-    "homepage": "http://picocms.org/",
+    "homepage": "https://picocms.org/",
     "license": "MIT",
     "authors": [
         {
@@ -18,7 +18,7 @@
         },
         {
             "name": "The Pico Community",
-            "homepage": "http://picocms.org/"
+            "homepage": "https://picocms.org/"
         },
         {
             "name": "Contributors",
@@ -26,17 +26,17 @@
         }
     ],
     "support": {
-        "docs": "http://picocms.org/docs",
+        "docs": "https://picocms.org/docs",
         "issues": "https://github.com/picocms/Pico/issues",
         "source": "https://github.com/picocms/Pico"
     },
     "require": {
-        "php": ">=5.3.6",
+        "php": ">=7.2.5",
         "ext-mbstring": "*",
-        "twig/twig": "^1.36",
-        "symfony/yaml" : "^2.8",
-        "erusev/parsedown": "1.8.0-beta-7",
-        "erusev/parsedown-extra": "0.8.0-beta-1"
+        "twig/twig": "^3.3.8",
+        "symfony/yaml" : "^5.4.3",
+        "erusev/parsedown": "1.7.4",
+        "erusev/parsedown-extra": "0.8.1"
     },
     "suggest": {
         "picocms/pico-theme": "Pico requires a theme to actually display the contents of your website. This is Pico's official default theme.",
@@ -44,13 +44,14 @@
         "picocms/composer-installer": "This Composer plugin is responsible for installing Pico plugins and themes using the Composer package manager."
     },
     "autoload": {
-        "psr-0": {
-            "Pico": "lib/",
-            "PicoPluginInterface": "lib/",
-            "AbstractPicoPlugin": "lib/"
+        "psr-4": {
+            "": "lib/"
         }
     },
     "extra": {
+        "allow-plugins": {
+            "picocms/*": true
+        },
         "branch-alias": {
             "dev-master": "2.1.x-dev",
             "dev-pico-3.0": "3.0.x-dev"
