@@ -10,18 +10,20 @@
  * License-Filename: LICENSE
  */
 
+declare(strict_types=1);
+
 /**
  * Abstract class to extend from when implementing a Pico plugin
  *
- * Please refer to {@see PicoPluginInterface} for more information about how
- * to develop a plugin for Pico.
+ * Please refer to {@see PicoPluginInterface} for more information about how to
+ * develop a plugin for Pico.
  *
  * @see PicoPluginInterface
  *
  * @author  Daniel Rudolf
- * @link    http://picocms.org
- * @license http://opensource.org/licenses/MIT The MIT License
- * @version 2.1
+ * @link    https://picocms.org
+ * @license https://opensource.org/licenses/MIT The MIT License
+ * @version 3.0
  */
 abstract class AbstractPicoPlugin implements PicoPluginInterface
 {
@@ -65,7 +67,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
      * @see PicoPluginInterface::getDependencies()
      * @var string[]
      */
-    protected $dependsOn = array();
+    protected $dependsOn = [];
 
     /**
      * List of plugin which depend on this plugin
@@ -98,7 +100,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
 
         if ($this->isEnabled() || ($eventName === 'onPluginsLoaded')) {
             if (method_exists($this, $eventName)) {
-                call_user_func_array(array($this, $eventName), $params);
+                call_user_func_array([ $this, $eventName ], $params);
             }
         }
     }
@@ -108,7 +110,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
      */
     protected function configEnabled()
     {
-        $pluginEnabled = $this->getPico()->getConfig(get_called_class() . '.enabled');
+        $pluginEnabled = $this->getPico()->getConfig(static::class . '.enabled');
         if ($pluginEnabled !== null) {
             $this->setEnabled($pluginEnabled);
         } else {
@@ -184,7 +186,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
      */
     public function getPluginConfig($configName = null, $default = null)
     {
-        $pluginConfig = $this->getPico()->getConfig(get_called_class(), array());
+        $pluginConfig = $this->getPico()->getConfig(static::class, []);
 
         if ($configName === null) {
             return $pluginConfig;
@@ -213,7 +215,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
 
         throw new BadMethodCallException(
             'Call to undefined method ' . get_class($this->getPico()) . '::' . $methodName . '() '
-            . 'through ' . get_called_class() . '::__call()'
+            . 'through ' . static::class . '::__call()'
         );
     }
 
@@ -233,7 +235,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
                 $plugin = $this->getPico()->getPlugin($pluginName);
             } catch (RuntimeException $e) {
                 throw new RuntimeException(
-                    "Unable to enable plugin '" . get_called_class() . "': "
+                    "Unable to enable plugin '" . static::class . "': "
                     . "Required plugin '" . $pluginName . "' not found"
                 );
             }
@@ -245,13 +247,13 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
                         $plugin->setEnabled(true, true, true);
                     } else {
                         throw new RuntimeException(
-                            "Unable to enable plugin '" . get_called_class() . "': "
+                            "Unable to enable plugin '" . static::class . "': "
                             . "Required plugin '" . $pluginName . "' was disabled manually"
                         );
                     }
                 } else {
                     throw new RuntimeException(
-                        "Unable to enable plugin '" . get_called_class() . "': "
+                        "Unable to enable plugin '" . static::class . "': "
                         . "Required plugin '" . $pluginName . "' is disabled"
                     );
                 }
@@ -287,7 +289,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
                             $plugin->setEnabled(false, true, true);
                         } else {
                             throw new RuntimeException(
-                                "Unable to disable plugin '" . get_called_class() . "': "
+                                "Unable to disable plugin '" . static::class . "': "
                                 . "Required by manually enabled plugin '" . $pluginName . "'"
                             );
                         }
@@ -297,7 +299,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
                 $dependantsList = 'plugin' . ((count($dependants) > 1) ? 's' : '') . ' '
                     . "'" . implode("', '", array_keys($dependants)) . "'";
                 throw new RuntimeException(
-                    "Unable to disable plugin '" . get_called_class() . "': "
+                    "Unable to disable plugin '" . static::class . "': "
                     . "Required by " . $dependantsList
                 );
             }
@@ -310,12 +312,12 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
     public function getDependants()
     {
         if ($this->dependants === null) {
-            $this->dependants = array();
+            $this->dependants = [];
             foreach ($this->getPico()->getPlugins() as $pluginName => $plugin) {
                 // only plugins which implement PicoPluginInterface support dependencies
                 if ($plugin instanceof PicoPluginInterface) {
                     $dependencies = $plugin->getDependencies();
-                    if (in_array(get_called_class(), $dependencies)) {
+                    if (in_array(static::class, $dependencies)) {
                         $this->dependants[$pluginName] = $plugin;
                     }
                 }
@@ -349,7 +351,7 @@ abstract class AbstractPicoPlugin implements PicoPluginInterface
 
             if (!$this->nativePlugin && ($pluginApiVersion > $picoApiVersion)) {
                 throw new RuntimeException(
-                    "Unable to enable plugin '" . get_called_class() . "': The plugin's API (version "
+                    "Unable to enable plugin '" . static::class . "': The plugin's API (version "
                     . $pluginApiVersion . ") isn't compatible with Pico's API (version " . $picoApiVersion . ")"
                 );
             }
